Add resumable streams, stream intercept limits (#29)
At the feature level, this PR adds the following:

* The ability to limit the number of concurrently intercepted streams.
When the number of matching streams exceeds the limit, those streams in
excess of the limit will be queued, and then intercepted when another
intercepted stream closes.
* This enables bardcast to make use of `Stream::set_monitor_stream()`
when using the monitor intercept mode with a limit of 1. This allows
audio from a specific application to be intercepted without needing to
set up virtual sinks in PulseAudio.
* The ability for audio consumers to detect when audio streams are
paused/resumed. This allows bardcast to remove its "speaking" state when
there is nothing actually playing audio, and restart the stream when
audio starts playing again.

A large refactoring was performed at the code level to make this
possible:

* Audio consumers (`discord`, `wav`) are now injected into the `snd`
module, which avoids the need to pass trait objects between the two.
* `Interceptor`s now own the PulseAudio `Stream` object, and exercise a
greater degree of control over its lifecycle:
* The `Stream` object is initialized lazily, upon the first intercept.
* The `Stream` object is dropped when the last application stream is
removed from the intercept.
* When all intercepted application streams are corked, bardcast will
cork its recording stream. This prevents silence from being generated by
PulseAudio, and allows audio consumers to react to true silence.
* There is now a single entry point for the `intercept` module, which
will determine the appropriate `Interceptor` implementation and start
the task to manage it.
* The general interface types exported by the `snd` module have been
further fleshed out. There is also an adapter that provides a `Read`
interface to types that implement `AsyncRead`, which has a customizable
read timeout.
* To allow the Discord consumer to reclaim the stream object when it
goes dormant, the `Lessor` and `Lease` types were added to the `util`
module, which allow calling code to return an owned object to a `Lessor`
when a `Lease` is dropped.

Closes #15.